install脚本修改

+++ .githooks/install.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-set -eu
+set -eux
 
 msg_color_error='\033[31m'
 msg_color_success='\033[32m'
@@ -22,11 +22,16 @@ function backup_hooks_scripts
     echo "${msg_color_warning}Begin backup_hooks_scripts ...${msg_color_none} \n"
     CUR_DIR=$(pwd)
     GIT_DIR=$(git rev-parse --show-toplevel)/.git
-    cd "${GIT_DIR}"
+    cd $GIT_DIR
 
-    set -e
+    if [ -e $GIT_DIR/hooks.old ]; then
+    rm -rf $GIT_DIR/hooks.old
+    fi
+
+    if [ -e $GIT_DIR/hooks ]; then
     mv hooks hooks.old
-    set +e
+    fi
+    
     mkdir hooks
     cd $CUR_DIR 
     echo "${msg_color_success}backup_hooks_scripts success ${msg_color_none} \n"
@@ -35,11 +40,12 @@ function backup_hooks_scripts
 function link_hooks_scripts
 {
     echo "${msg_color_warning}Begin link_hooks_scripts ...${msg_color_none} \n"
+    CUR_DIR=`pwd`
     HOOK_DIR=$(git rev-parse --show-toplevel)/.git/hooks
     for file in $1
     do
-        chmod +x "${file}"
-        ln -s "${file}" $HOOK_DIR/"${file}"
+        ln -s $CUR_DIR/$file $HOOK_DIR/$file
+        echo "${msg_color_success} ${file} link  to ${HOOK_DIR}/${file} ${msg_color_none} \n"
     done
     echo "${msg_color_success}link_hooks_scripts success ${msg_color_none} \n"
 }
